refactor!: libraryfication
cloudflare-ddns is now based on libcloudflare-ddns, a library that implements all the functionality needed to implement a simple client that updates DNS records. It can be used by external projects without issues, and provides both convenient methods and lower level ones. Thanks to the libraryfication, the codebase is now much cleaner and testable, and performance should be the same :)