refractor: remove one get_record_raw()
The version of get_record_raw() that managed its own cURL handle was
useless, as it didn't provide any performance boost compared to
get_record() and you also had to parse the response yourself. Not only
useless, but actively disturbing, because it was impossible to pass the
function to std::async without a long cast needed to disambiguate the
overloaded funciton.