fix(exe): properly handle IP v4 and v6 hosts
An oversight in the get_local_ip() logic could've led to errors when an
host was both IPv4 and IPv6 capable.

In particular, the function didn't enforce the IP version, and this made
it impossible to reliably update a given record type, as get_local_ip()
could return either IPv4 or IPv6 addresses.

get_local_ip() is now executed after get_record_raw(), and
get_local_ip() is forced to use IPv4 or IPv6 if the get_record() call
returns an A or AAAA record, respectively.

This has the side effect of eliminating parallelism opportunities, as
get_local_ip() can only be called after determining the record type with
get_record(). This makes cloudflare-ddns slower in theory, but this
unavoidable.

Thanks to Stephan Lachnit (@stephanlachnit) for reporting and helping!

Fixes #12